src/requirements.txt
requests
github3.py
slackclient